+++ src/Braintree/TransactionOptionsRequest.cs
@@ -18,6 +18,7 @@ namespace Braintree
         public TransactionOptionsPayPalRequest PayPal { get; set; }
         public TransactionOptionsThreeDSecureRequest ThreeDSecure { get; set; }
         public TransactionOptionsAmexRewardsRequest AmexRewards { get; set; }
+        public TransactionOptionsVenmoRequest Venmo { get; set; }
 
         public override string ToXml(string root)
         {
@@ -45,7 +46,8 @@ namespace Braintree
                 AddElement("skip-cvv", SkipCvv).
                 AddElement("three-d-secure", ThreeDSecure).
                 AddElement("paypal", PayPal).
-                AddElement("amex-rewards", AmexRewards);
+                AddElement("amex-rewards", AmexRewards).
+                AddElement("venmo", Venmo);
         }
     }
 }